add sentry project context correctly
