+++ src/main/resources/META-INF/plugin.xml
@@ -338,9 +338,6 @@
         <listener class="org.move.ide.notifications.UpdateNotificationsOnSettingsChangeListener"
                   topic="org.move.cli.settings.MoveSettingsListener"
                   activeInTestMode="false"/>
-        <listener
-            class="org.move.cli.sentryReporter.SentryContextsProjectsListener"
-            topic="org.move.cli.MoveProjectsService$MoveProjectsListener" />
     </projectListeners>
 
     <actions>
